requirements.txt
fastapi
uvicorn
python-dotenv
litellm
openai
requests
pydantic
sqlalchemy
aiohttp
httpx
tiktoken
langsmith
redis
psycopg2-binary
jinja2
python-multipart
